refactors migrations to server-common
