+++ apps/crn-server/scripts/migration.template
@@ -1,6 +1,6 @@
 /* istanbul ignore file */
 import { Results, Squidex } from '@asap-hub/squidex';
-import { Migration } from '../handlers/webhooks/webhook-run-migrations';
+import { Migration } from '@asap-hub/server-common';
 
 export default class MoveResearchOutputTextToDescription extends Migration {
   up = async (): Promise<void> => {};
