+++ apps/crn-server/src/utils/migrations.ts
@@ -1,45 +1,9 @@
-import {
-  Entity,
-  Rest,
-  Results,
-  SquidexRest,
-  SquidexRestClient,
-} from '@asap-hub/squidex';
+import { applyToAllItemsInCollectionFactory } from '@asap-hub/server-common';
 import { appName, baseUrl } from '../config';
 import { getAuthToken } from './auth';
 
-export const applyToAllItemsInCollection = async <
-  T extends Entity & Rest<unknown>,
->(
-  entityName: string,
-  processEntity: (
-    entity: T,
-    squidexClient: SquidexRestClient<T>,
-  ) => Promise<void>,
-): Promise<void> => {
-  const squidexClient = new SquidexRest<T>(
-    getAuthToken,
-    entityName,
-    { appName, baseUrl },
-    {
-      unpublished: true,
-    },
-  );
-
-  let pointer = 0;
-  let result: Results<T>;
-
-  do {
-    result = await squidexClient.fetch({
-      $top: 10,
-      $skip: pointer,
-      $orderby: 'created asc',
-    });
-
-    for (const item of result.items) {
-      await processEntity(item, squidexClient);
-    }
-
-    pointer += 10;
-  } while (pointer < result.total);
-};
+export const applyToAllItemsInCollection = applyToAllItemsInCollectionFactory(
+  appName,
+  baseUrl,
+  getAuthToken,
+);
