+++ packages/server-common/src/handlers/webhooks/index.ts
@@ -1,2 +1,3 @@
 export * from './connect-by-code.handler';
+export * from './run-migrations.handler';
 export * from './squidex.handler';
